Fixed few issues resolving package installability, described below
1) The Makefile is needed because the lcurses option is required when
using curses in a C file. Else, installing the package results in
complaints from rcurses.c not recognizing the curses symbols (e.g.
"echo").
2) The R_init_rcurses() function (namely, the R_registerRoutines() call,
which R_init_rcurses() does) should be in the rcurses.c because when the
rcurses dll is loaded, that's when R will look for the R_init_rcurses
function.